misc: do not handling sorting of tensors by name and do linear scan for now

+++ lib/safetensors_types.h
@@ -106,8 +106,8 @@ struct safetensors_tensor
     enum safetensors_dtype dtype;
     uint8_t rank;
     uint32_t *shape;
-    uint64_t start;
-    uint64_t end;
+    size_t start;
+    size_t end;
 };
 
 /// Holds the tensor definitions table
